[FIX] useInputFirst関数を再度作成(副作用切り出し)
- InSourceTestingを削除
- composablesの中にcomposablesを作成し副作用切り出し
- containerではなくcomposables側にテスト書き直し
